==> RecommenderSystem/2_Evaluation.ipynb <==
import os
import pandas as pd

from tqdm.notebook import tqdm as log_progress


eval_path = "Evaulation/Complex/F"

feature_path = "Processed/UserChannel"
feature_files = [f for f in os.listdir(feature_path) if os.path.isfile(os.path.join(feature_path, f))]

cluster_eval_files = []

cluster_top_result = [0, None]
for feature_file in feature_files:
    cl_eval_path = os.path.join(eval_path,feature_file)
    if(not os.path.exists(cl_eval_path)):
        continue
    cluster_eval_files = [f for f in os.listdir(cl_eval_path) if os.path.isfile(os.path.join(cl_eval_path, f))]
    for cluster_file in cluster_eval_files:
        pkl_file = os.path.join(cl_eval_path, cluster_file)
        if(not os.path.exists(pkl_file)):
            continue
        eval_file_df = pd.read_pickle(pkl_file)
        top_AUC=eval_file_df.groupby(["name"]).mean()['AUC'].sort_values(ascending=False).iloc[0]
        if(top_AUC > cluster_top_result[0]):
            cluster_top_result[0] = top_AUC
            cluster_top_result[1] = pkl_file

org_clustr_path = 'OrgUnitClustering.pkl'
org_cluster_eval_df = pd.read_pickle(org_clustr_path)

org_cluster_eval_desc_df = org_cluster_eval_df.groupby(["name"]).describe()
display(org_cluster_eval_desc_df)

team_cluster_path = 'TeamsClustering.pkl'
team_cluster_eval_df = pd.read_pickle(team_cluster_path)

team_cluster_eval_desc_df = team_cluster_eval_df.groupby(["name"]).describe()
display(team_cluster_eval_desc_df)

eval_path = "."
eval_files = [f for f in os.listdir(eval_path) if os.path.isfile(os.path.join(eval_path, f)) and f.endswith(".pkl") and f.startswith("user")]

top_result = [0, None]
for eval_file in log_progress(eval_files):    
    eval_file_df = pd.read_pickle(eval_file).sort_values(["AUC"], ascending=False)
    top_AUC = eval_file_df.iloc[0].AUC    
    if(top_AUC > top_result[0]):
        top_result[0] = top_AUC
        top_result[1] = eval_file

eval_path = "Evaulation/Complex/F"
feature_path = "Processed/UserChannel"
feature_files = [f for f in os.listdir(feature_path) if os.path.isfile(os.path.join(feature_path, f))]

cluster_eval_files = [f for f in os.listdir(eval_path) if os.path.isfile(os.path.join(eval_path, f)) and f.endswith(".pkl") and f.startswith("clusters")]

cluster_top_result = [0, None]
for eval_file in log_progress(cluster_eval_files):    
    eval_file_df = pd.read_pickle(eval_file).groupby(["name"])
    top_AUC = eval_file_df.describe()['AUC'].sort_values(["mean"], ascending=False).iloc[0]['mean']         
    if(top_AUC > cluster_top_result[0]):
        cluster_top_result[0] = top_AUC
        cluster_top_result[1] = eval_file

eval_path = "Evaulation/Complex/Ego/"


eval_files = [os.path.join(eval_path, f) for f in os.listdir(eval_path) if os.path.isfile(os.path.join(eval_path, f))]
eval_files.extend([os.path.join("Evaulation/Complex/", f) for f in os.listdir("Evaulation/Complex") if os.path.isfile(os.path.join("Evaulation/Complex", f))])
top_result = [0, None]
for eval_file in log_progress(eval_files):    

    eval_file_df = pd.read_pickle(eval_file).sort_values(["AUC"], ascending=False)
    top_AUC = eval_file_df.iloc[0].AUC    

    if(top_AUC > top_result[0]):
        top_result[0] = top_AUC
        top_result[1] = eval_file

display(top_result)

eval_path = "Evaulation/Simple/"


eval_files = [os.path.join(eval_path, f) for f in os.listdir(eval_path) if os.path.isfile(os.path.join(eval_path, f))]
top_result = [0, None]
for eval_file in log_progress(eval_files):    

    eval_file_df = pd.read_pickle(eval_file).sort_values(["AUC"], ascending=False)
    top_AUC = eval_file_df.iloc[0].AUC    

    if(top_AUC > top_result[0]):
        top_result[0] = top_AUC
        top_result[1] = eval_file

display(top_result)
