==> RecommenderSystem/3_SimpleMeasure.ipynb <==
import random
import pandas as pd
from cdhf.data import Data
from utils.helpers import execute, save_pckl, normalize, create_init_dataframe

pd.options.mode.chained_assignment = None  # default='warn'

random.seed(42)
random_sample = random.sample(range(10, 300), 1)
data = Data("../input/mmdata.json")
data.load_all()

df = create_init_dataframe(data)

results = execute(df, random_sample)
path = "Evaulation/Simple/AllData.pckl"
save_pckl(path, results)

df["score"] = normalize(df["score"]) * 10000
results = execute(df, random_sample)
path = "Evaulation/Simple/AllData-Normalized.pckl"
save_pckl(path, results)
display(results)
